Added tests and edited documentation
