+++ docs/content/en/integrations/import.md
@@ -105,6 +105,65 @@ processed and made available in the \'Finding View\' page.
 
 Import HTML reports from Burp Enterprise Edition
 
+### Burp GraphQL
+
+Import the JSON data returned from the BurpSuite Enterprise GraphQL API. Append all the
+issues returned to a list and save it as the value for the key "Issues". There is no need
+to filter duplicates, the parser will automatically combine issues with the same name.
+
+Example:
+
+```JSON
+{
+    "Issues": [
+        {
+            "issue_type": {
+                "name": "Cross-site scripting (reflected)",
+                "description_html": "Issue Description",
+                "remediation_html": "Issue Remediation",
+                "vulnerability_classifications_html": "<li><a href=\"https://cwe.mitre.org/data/definitions/79.html\">CWE-79: Improper Neutralization of Input During Web Page Generation ('Cross-site Scripting')</a></li>",
+                "references_html": "<li><a href=\"https://portswigger.net/web-security/cross-site-scripting\">Cross-site scripting</a></li>"
+            },
+            "description_html": "Details",
+            "remediation_html": "Remediation Details",
+            "severity": "high",
+            "path": "/path",
+            "origin": "https://www.example.com",
+            "evidence": [
+                {
+                    "request_index": 0,
+                    "request_segments": [
+                        {
+                            "data_html": "GET"
+                        },
+                        {
+                            "highlight_html": "data"
+                        },
+                        {
+                            "data_html": " HTTP More data"
+                        }
+                    ]
+                },
+                {
+                    "response_index": 0,
+                    "response_segments": [
+                        {
+                            "data_html": "HTTP/2 200 OK "
+                        },
+                        {
+                            "highlight_html": "data"
+                        },
+                        {
+                            "data_html": "More data"
+                        }
+                    ]
+                }
+            ]
+        }
+    ]
+}
+```
+
 ### CargoAudit Scan
 
 Import JSON output of cargo-audit scan report <https://crates.io/crates/cargo-audit>
