+++ dojo/unittests/scans/burp_graphql/no_findings.json
@@ -1,0 +1,3 @@
+{
+    "Issues": []
+}
